fix: dcl status with negative first coord
